refactor(config): split config into domain modules
Move the monolithic config schema into focused modules under src/config
while preserving crate::config re-exports and load/save behavior.

Split config tests by domain so the new module tree stays within the
400-line guardrail.

Closes #400
